final_project/requirements.txt
requests
pandas
kafka-python
python-dotenv
